move parser tests to respective packages
