+++ src/test/java/seedu/address/logic/parser/event/EaddCommandParserTest.java
@@ -1,0 +1,12 @@
+package seedu.address.logic.parser.event;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class EaddCommandParserTest {
+
+    @Test
+    void parse() {
+    }
+}
